# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.LoginInput:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
...
  models.RefreshInput:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  models.Song:
    properties:
      album_id:
        description: 加索引，提高查询速度
        type: integer
      created_at:
        type: string
      duration:
        description: 单位：秒
        type: integer
      id:
        type: integer
      title:
        type: string
      track_number:
        description: 专辑内排序
        type: integer
      updated_at:
        type: string
    type: object
  models.SongForm:
    properties:
      album_id:
        type: integer
      duration:
        description: 单位：秒
        minimum: 0
        type: integer
      title:
        maxLength: 100
        minLength: 1
        type: string
      track_number:
        description: 专辑内排序
        minimum: 1
        type: integer
    required:
    - album_id
    - duration
    - title
    - track_number
    type: object
  models.Team:
    properties:
      champions:
        type: integer
      city:
        type: string
      created_at:
        type: string
      divide:
        type: string
      id:
        type: integer
      logo:
        type: string
      name:
        type: string
      part:
        type: string
      updated_at:
        type: string
    type: object
  utils.ApiRes:
    properties:
      code:
        type: integer
      data: {}
      errors: {}
      message:
        type: string
      timestamp:
        type: string
    type: object
  utils.BusinessError:
    properties:
      bizCode:
        type: integer
      details: {}
      err: {}
      httpCode:
        type: integer
    type: object
  utils.ListResponse:
    properties:
      list: {}
      total:
        type: integer
    type: object
info:
  contact: {}
  description: This is a sample go-learn server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: go-learn API
  version: "1.0"
paths:
...
  /albums/{id}/songs:
    get:
      description: 根据专辑ID获取该专辑下的所有歌曲，支持分页
      parameters:
      - description: 专辑ID
        in: path
        name: id
        required: true
        type: integer
      - description: 每页数量
        in: query
        name: limit
        type: integer
      - description: 偏移量
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ListResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.BusinessError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.BusinessError'
      summary: 获取专辑下的所有歌曲
      tags:
      - songs
...
  /login:
    post:
      consumes:
      - application/json
      description: 用户登录，获取access_token和refresh_token
      parameters:
      - description: 登录信息
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/models.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.BusinessError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.BusinessError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.BusinessError'
      summary: 用户登录
      tags:
      - users
...
  /refresh:
    post:
      consumes:
      - application/json
      description: 使用refresh_token刷新access_token
      parameters:
      - description: 刷新信息
        in: body
        name: refresh
        required: true
        schema:
          $ref: '#/definitions/models.RefreshInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.BusinessError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.BusinessError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.BusinessError'
      summary: 刷新Token
      tags:
      - users
  /register:
    post:
      consumes:
      - application/json
      description: 用户注册
      parameters:
      - description: 注册信息
        in: body
        name: register
        required: true
        schema:
          $ref: '#/definitions/models.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.BusinessError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.BusinessError'
      summary: 用户注册
      tags:
      - users
  /songs:
    get:
      description: 获取所有歌曲，支持分页
      parameters:
      - description: 每页数量
        in: query
        name: limit
        type: integer
      - description: 偏移量
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.BusinessError'
      summary: 获取所有歌曲
      tags:
      - songs
    post:
      consumes:
      - application/json
      description: 创建一首新歌
      parameters:
      - description: 歌曲信息
        in: body
        name: song
        required: true
        schema:
          $ref: '#/definitions/models.SongForm'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Song'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.BusinessError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.BusinessError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.BusinessError'
      summary: 创建歌曲
      tags:
      - songs
  /songs/{id}:
    delete:
      description: 根据ID删除歌曲
      parameters:
      - description: 歌曲ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.BusinessError'
      summary: 删除歌曲
      tags:
      - songs
    get:
      description: 根据ID获取歌曲详情
      parameters:
      - description: 歌曲ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Song'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.BusinessError'
      summary: 获取单个歌曲
      tags:
      - songs
    put:
      consumes:
      - application/json
      description: 根据ID更新歌曲信息
      parameters:
      - description: 歌曲ID
        in: path
        name: id
        required: true
        type: integer
      - description: 歌曲信息
        in: body
        name: song
        required: true
        schema:
          $ref: '#/definitions/models.SongForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.BusinessError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.BusinessError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.BusinessError'
      summary: 更新歌曲
      tags:
      - songs
  /teams:
    get:
      description: 获取所有团队，支持分页
      parameters:
      - description: 每页数量
        in: query
        name: limit
        type: integer
      - description: 偏移量
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ApiRes'
      summary: 获取所有团队
      tags:
      - teams
  /teams/{id}:
    get:
      description: 根据ID获取团队详情
      parameters:
      - description: 团队ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Team'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.BusinessError'
      summary: 获取单个团队
      tags:
      - teams
swagger: "2.0"